% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_instances}
\alias{lightsail_get_instances}
\title{Returns information about all Amazon Lightsail virtual private servers,
or \emph{instances}}
\usage{
lightsail_get_instances(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get
instances request.}
}
\description{
Returns information about all Amazon Lightsail virtual private servers,
or \emph{instances}.
}
\section{Request syntax}{
\preformatted{lightsail$get_instances(
  pageToken = "string"
)
}
}

\keyword{internal}
