% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_peer_vpc}
\alias{lightsail_peer_vpc}
\title{Tries to peer the Lightsail VPC with the user's default VPC}
\usage{
lightsail_peer_vpc()
}
\description{
Tries to peer the Lightsail VPC with the user's default VPC.
}
\section{Request syntax}{
\preformatted{lightsail$peer_vpc()
}
}

\keyword{internal}
